Starting game
The tetris shapes are randomised and are drawn correctly to the screen. All rotation stages are correct. Tetris shapes stop when they hit the bottom cell. (Need to add = On keypress rotate shapes / if shapes exceed the height of the cells game over / check when a line is created then remove them and add score)